requirements.txt
flask
flask-cors
python-dotenv
flask-sqlalchemy
sqlalchemy
psycopg2-binary